kdebug : just a constant change kappdcopiface : needed change so that dcopidl2cpp can work (ulong no more supported) kapplication : start of port, still a lot to do ;)
svn path=/branches/work/kde4/kdelibs/; revision=411694

+++ kdecore/kappdcopiface.h
@@ -63,7 +63,7 @@ k_dcop:
 
 	void reparseConfiguration();
 
-	void updateUserTimestamp( ulong time );
+	void updateUserTimestamp( quint32 time );
 
 private:
 	KApplication *m_KApplication;
